Refuctoring
remove sync folder.
