--- Src/ClArc.Tests/Async/InputData.cs
@@ -1,8 +1,0 @@
-﻿using ClArc.Async.Core;
-
-namespace ClArc.Tests.Async
-{
-    public class InputData : IInputData<OutputData>
-    {
-    }
-}
