--- Src/ClArc.Tests/Async/OutputData.cs
@@ -1,11 +1,0 @@
-﻿using ClArc.Async.Core;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ClArc.Tests.Async
-{
-    public class OutputData : IOutputDataAsync
-    {
-    }
-}
