--- Src/ClArc.Tests/Async/ThrowsExceptionInteractor.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-using ClArc.Async.Core;
-
-namespace ClArc.Tests.Async
-{
-    public class ThrowsExceptionInteractor : IInputPort<InputData, OutputData, Task<OutputData>>
-    {
-
-
-        public  Task<OutputData> Handle(InputData request)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
