+++ Src/ClArc.Tests/Sync/AsyncInteractor.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 using System.Threading.Tasks;
-using ClArc.Sync.Core;
+using ClArc.Core;
 
 namespace ClArc.Tests.Sync
 {
