+++ Src/ClArc.Tests/Sync/ThrowsExceptionInteractor.cs
@@ -1,5 +1,5 @@
 ﻿using System;
-using ClArc.Sync.Core;
+using ClArc.Core;
 
 namespace ClArc.Tests.Sync
 {
