--- Src/ClArc/Async/Core/IInputData.cs
@@ -1,6 +1,0 @@
-﻿namespace ClArc.Async.Core
-{
-    public interface IInputData<out TOutputDataAsync> where TOutputDataAsync : IOutputDataAsync
-    {
-    }
-}
