--- Src/ClArc/Async/Invoker/IUseCaseInvoker.cs
@@ -1,10 +1,0 @@
-﻿using ClArc.Async.Core;
-using System.Threading.Tasks;
-
-namespace ClArc.Async.Invoker
-{
-    public interface IUseCaseInvokerAsync
-    {
-        Task<TResponse> Invoke<TResponse>(IInputData<TResponse> inputData) where TResponse : IOutputDataAsync;
-    }
-}
