--- Src/ClArc/Async/Invoker/IUseCaseInvokerFactory.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace ClArc.Async.Invoker
-{
-    public interface IUseCaseInvokerFactoryAsync
-    {
-        IUseCaseInvokerAsync Generate(Type usecaseType, Type implementsType, IServiceProvider provider);
-    }
-}
