--- Src/ClArc/Async/Invoker/UseCaseInvokerDefaultFactory.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace ClArc.Async.Invoker
-{
-    public class UseCaseInvokerDefaultFactoryAsync :     IUseCaseInvokerFactoryAsync
-    {
-        public IUseCaseInvokerAsync Generate(Type usecaseType, Type implementsType, IServiceProvider provider)
-        {
-            return new UseCaseInvoker(usecaseType, implementsType, provider);
-        }
-    }
-}
