+++ Src/ClArc/Core/IInputData.cs
@@ -1,0 +1,6 @@
+﻿namespace ClArc.Core
+{
+    public interface IInputData<out TOutputData> where TOutputData : IOutputData
+    {
+    }
+}
