+++ Src/ClArc/Core/IInputPortVoidOutput.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace ClArc.Sync.Core
+namespace ClArc.Core
 {
     public interface IInputPortVoidOutput<in TInputDataVoidOutput>
         where TInputDataVoidOutput : IInputDataVoidOutput
