+++ Src/ClArc/Core/IOutputData.cs
@@ -1,0 +1,6 @@
+﻿namespace ClArc.Core
+{
+    public interface IOutputData
+    {
+    }
+}
