+++ Src/ClArc/Invoker/IUseCaseInvokerFactory.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace ClArc.Invoker
+{
+    public interface IUseCaseInvokerFactory
+    {
+        IUseCaseInvoker Generate(Type usecaseType, Type implementsType, IServiceProvider provider);
+    }
+}
