--- Src/ClArc/Sync/Core/EmptyOutputData.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace ClArc.Sync.Core
-{
-    public class EmptyOutputData : IOutputData
-    {
-    }
-}
