--- Src/ClArc/Sync/Core/IInputData.cs
@@ -1,6 +1,0 @@
-﻿namespace ClArc.Sync.Core
-{
-    public interface IInputData<out TOutputData> where TOutputData : IOutputData
-    {
-    }
-}
