--- Src/ClArc/Sync/Core/IOutputData.cs
@@ -1,6 +1,0 @@
-﻿namespace ClArc.Sync.Core
-{
-    public interface IOutputData
-    {
-    }
-}
