--- Src/ClArc/Sync/Invoker/IUseCaseInvokerFactory.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace ClArc.Sync.Invoker
-{
-    public interface IUseCaseInvokerFactory
-    {
-        IUseCaseInvoker Generate(Type usecaseType, Type implementsType, IServiceProvider provider);
-    }
-}
