--- Src/ClArc/Sync/Invoker/UseCaseInvokerDefaultFactory.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace ClArc.Sync.Invoker
-{
-    public class UseCaseInvokerDefaultFactory : IUseCaseInvokerFactory
-    {
-        public IUseCaseInvoker Generate(Type usecaseType, Type implementsType, IServiceProvider provider)
-        {
-            return new UseCaseInvoker(usecaseType, implementsType, provider);
-        }
-    }
-}
